Update txflow.rst
OLD: 
The chaincode (containing a set of key value pairs representing the initial
state of the radish market) is installed on the peers and instantiated on the
channel. 
NEW:
The chaincode is installed on the peers and instantiated on the
channel. 
EXPLANATION: I might be wrong but the chaincode does not include the status of a key value . The status of key value is stored in the world state (or sideDB) or the ledger.